+++ native/CMakeLists.txt
@@ -1422,6 +1422,8 @@ cocos_source_files(
                  cocos/renderer/pipeline/custom/NativeBuiltinUtils.cpp
                  cocos/renderer/pipeline/custom/NativeBuiltinUtils.h
                  cocos/renderer/pipeline/custom/NativeExecutor.cpp
+                 cocos/renderer/pipeline/custom/NativeExecutorDescriptor.cpp
+                 cocos/renderer/pipeline/custom/NativeExecutorRenderGraph.h
                  cocos/renderer/pipeline/custom/NativeFactory.cpp
                  cocos/renderer/pipeline/custom/NativeFwd.h
                  cocos/renderer/pipeline/custom/NativePipeline.cpp
